Refactor decoder object in src/services and src/typings.
